3_Vegetable_Classification/Deployment/requirements.txt
streamlit
requests
Pillow
tensorflow-cpu
numpy